requirements.txt
crewai
langchain-openai
python-dotenv
openai
xgboost
